[mod] #116 인자 nullable 수정

+++ core/model/src/main/java/org/sopt/model/category/CategoryChangePriority.kt
@@ -2,5 +2,5 @@ package org.sopt.model.category
 
 data class CategoryChangePriority(
   val categoryId: Long,
-  val newPriority: Long,
+  val newPriority: Int,
 )
